pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sfm"
version = "1.0.0"
description = "Public submission code for the Screening Foundation Model (SFM)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
  { name = "Longzhen Yang" },
  { name = "Yihang Liu" },
  { name = "Juzhao Zhang" },
  { name = "Ying Wen" },
  { name = "Jiaxiong Yang" },
  { name = "Ziteng Yang" },
  { name = "Qi Liu" },
  { name = "Lina Lu" },
  { name = "Yi Xu" },
  { name = "Danli Shi", email = "danli.shi@polyu.edu.hk" },
  { name = "Mingguang He", email = "mingguang.he@polyu.edu.hk" },
  { name = "Haidong Zou" },
  { name = "Ching Yu Cheng", email = "chingyu.cheng@nus.edu.sg" },
  { name = "Jide He", email = "jide_he@fudan.edu.cn" },
  { name = "Senlin Lin", email = "woodylin@yeah.net" },
  { name = "Lianghua He", email = "helianghua@tongji.edu.cn" },
]
dependencies = [
  "numpy==2.1.2",
  "pandas==2.2.3",
  "scipy==1.14.1",
  "scikit-learn==1.7.2",
  "Pillow==11.0.0",
  "PyYAML==6.0.3",
  "openpyxl==3.1.5",
  "lifelines==0.30.0",
  "huggingface-hub==1.3.5",
]

[project.optional-dependencies]
torch = [
  "torch==2.5.1",
  "torchvision==0.20.1",
  "timm==1.0.15",
]
multimodal = [
  "transformers==5.0.0",
  "sentencepiece==0.2.0",
]
moe = [
  "tutel @ git+https://github.com/microsoft/Tutel.git",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sfm*", "analysis*"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
